# biblioteca-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticación",
        "description": "Controlador para registro, login y autenticación con JWT",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registro de usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado con éxito"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "juan@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token JWT retornado"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener datos del usuario autenticado",
                "operationId": "a611d63f247c572e29d9d9a4720b5562",
                "responses": {
                    "200": {
                        "description": "Datos del usuario"
                    },
                    "401": {
                        "description": "Token inválido o ausente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "Token inválido o ausente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/libros": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Obtener todos los libros",
                "operationId": "08f70de053dbd61674113fd6369fe329",
                "responses": {
                    "200": {
                        "description": "Lista de libros"
                    }
                }
            },
            "post": {
                "tags": [
                    "Libros"
                ],
                "summary": "Crear un nuevo libro",
                "operationId": "c4bb407519a52e2207d4fe944d74b80e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "autor",
                                    "genero"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Cien años de soledad"
                                    },
                                    "autor": {
                                        "type": "string",
                                        "example": "Gabriel García Márquez"
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "Ficción"
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Libro creado exitosamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/libros/{id}": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Obtener detalles de un libro",
                "operationId": "21bd0cf3073a238612092e6f7a7b2162",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del libro"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Libros"
                ],
                "summary": "Actualizar un libro",
                "operationId": "28deef916fa3679afeb579b4f14bc800",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Nuevo título"
                                    },
                                    "autor": {
                                        "type": "string",
                                        "example": "Nuevo autor"
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "Ciencia ficción"
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Libro actualizado"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Libros"
                ],
                "summary": "Eliminar un libro",
                "operationId": "ad6977a2ef1e3093648c03b2ce5c1814",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Libro eliminado"
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}